Fix DRB345 report parsing
Whether or not to report zero, one or two DRB345 alleles is not simply
deterministic. It is based on the acutal DRB1 haplotype.

We need to look at source data and actually add the appropriate DRB345
alleles if and only if necessary.

This also means the correct data structure for these types is a list.