Update haplotype database
Update the haplotypes to 2013 values, and replace DRB1-DQB1 with
DRB345-DRB1-DQB1 (still missing the DQA's but better than without
DRB345).

The 2013 data uses more specific race codes instead of general race
groups, so we have to map the codes to the groups (as DonorNet still
uses the race groups).