Rewrite haplotype scoring to improve performance
We prioritize haplotype sets on the CWD status of their alleles, and
the population frequencies of the haplotype. Priority is given to CWD
status - which is relatively quick to compute. Population frequency can
take longer, and really only needs to be looked up once per haplotype
per ethnicity.

Previously we were computing a combined score using both of these
factors for ALL resulting haplotypes. This was taking too long on the
older fairview VMs.

By first binning the haplotype sets by CWD scores, then computing the
frequency scores of only the top bin (and caching the scores of each
individual haplotype), we have improved performance 10x.