Move unknown hap threshold to frequency parsing
Haplotypes with a frequency of "0" in the haplotype table are
technically entered as "0.00000000000000", which as a double is greater
than "0" and thus does not trigger the missing haplotype penalty.

During parsing, we need to check this value and set haplotypes with a
frequency below the cutoff to an explicit "0" frequency.