Refactor SureTyper DRB345 parsing
The DRB345 format is significantly different between old and new
versions of SureTyper. In one the allele data is mixed into the
DRB1 section, and contains a consistent allele start token (locus with
first field). In the other, the DRB345's are split to their own
section, but the allele start token is in a different form (locus-only).

The general logic imlemented here should cover both cases.