Fix two bugs in batch test validation
Bug-1 occurred when DonorCheck is first run, a test is added to the
validation, and then test is opened or run before DonorCheck is closed.

Bug-2 occurred when a test with the same input file for both inputs is
added as a test. In this instance, the test file is now copied a second
time with an altered name.